2022_06_29_Zapata_H2_trotter/requirements.txt
backports.cached-property==1.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cirq-core==0.14.1
cirq-google==0.14.1
cryptography==36.0.2
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.4
duet==0.2.5
fonttools==4.32.0
google-api-core==1.31.5
google-auth==1.35.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
kiwisolver==1.4.2
lea==3.4.2
matplotlib==3.5.1
mpmath==1.2.1
networkx==2.4
ntlm-auth==1.5.0
numpy==1.21.5
openfermion==1.3.0
overrides==3.1.0
packaging==21.3
pandas==1.3.5
pbr==5.8.1
Pillow==9.1.0
ply==3.11
protobuf==3.20.0
psutil==5.9.0
PubChemPy==1.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.8
python-constraint==1.4.0
python-dateutil==2.8.2
python-rapidjson==1.6
pytz==2022.1
qe-qiskit @ file:///Users/michalstechly/zapata_projects/darpa/darpa-circuits/2022_04_11_zapata_toy_trotter/qe-qiskit
qiskit==0.33.1
qiskit-aer==0.9.1
qiskit-ibmq-provider==0.18.2
qiskit-ignis==0.7.0
qiskit-terra==0.19.1
requests==2.27.1
requests-ntlm==1.1.0
retworkx==0.11.0
rsa==4.8
scipy==1.7.3
six==1.16.0
sortedcontainers==2.4.0
stevedore==3.5.0
symengine==0.9.2
sympy==1.9
tqdm==4.64.0
tweedledum==1.1.1
typing-extensions==3.10.0.0
urllib3==1.26.9
websocket-client==1.3.2
z-quantum-core @ file:///Users/michalstechly/zapata_projects/darpa/darpa-circuits/2022_04_11_zapata_toy_trotter/z-quantum-core
zipp==3.8.0
